requirements.txt

# Requires Python =3.10.16

# Web scraping
pyppeteer>=2.0.0
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.60.0
anthropic>=0.43.0
python-dotenv>=1.0.0
google-generativeai>=0.6.0 

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.70.0

# Data processing and visualization
yfinance>=0.2.36
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.1

# Tabulate for pretty-printing tables
tabulate

# Utilities
aiohttp==3.10.11
requests>=2.32.0
uuid
